Make scent system better

+++ src/core/game.rs
@@ -119,7 +119,7 @@ impl Game {
             
           }
 
-          if !self.world.is_valid(self.world.player.pos.x as usize, self.world.player.pos.y as usize) {
+          if !self.world.is_valid(self.world.player.pos.x, self.world.player.pos.y) {
             self.world.player.pos = oldpos;
             self.state = State::Act(Actions::Unknown);
           }
